% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{export_data}
\alias{export_data}
\title{Better CSV Exporter}
\usage{
export_data(data, filename)
}
\arguments{
\item{data}{A data frame to export.}

\item{filename}{Character string for the output path.}
}
\description{
A modernized version of legacy export functions.
}
\examples{
\dontrun{
export_data(mtcars, "my_data.csv")
}
}
